chore: Refactor package structure
