+++ src/main/java/se/icus/mag/modsettings/Main.java
@@ -8,7 +8,7 @@ import net.minecraft.client.util.InputUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
-import se.icus.mag.modsettings.gui.ModSettingsScreen;
+import se.icus.mag.modsettings.gui.screen.ModSettingsScreen;
 
 public class Main implements ClientModInitializer {
     public static final Logger LOGGER = LogManager.getLogger("modsettings");
